+++ pkcs11/src/backend/db/object.rs
@@ -331,11 +331,9 @@ pub fn from_key_data(
 
     let key_attrs = match key_data.r#type {
         KeyType::Rsa => configure_rsa(&key_data)?,
-        KeyType::Curve25519
-        | KeyType::EcP224
-        | KeyType::EcP256
-        | KeyType::EcP384
-        | KeyType::EcP521 => configure_ec(&key_data)?,
+        KeyType::Curve25519 | KeyType::EcP256 | KeyType::EcP384 | KeyType::EcP521 => {
+            configure_ec(&key_data)?
+        }
         KeyType::Generic => configure_generic()?,
     };
     attrs.extend(key_attrs.attrs);